Move function inside browser api
